+++ src/Configuration.cpp
@@ -95,7 +95,11 @@ nlohmann::basic_json<> Configuration::json() const {
 
   auto& fun = config["functionality"];
   fun["trace_threshold"] = functionality.traceThreshold;
+  fun["approximate_checking_threshold"] =
+      functionality.approximateCheckingThreshold;
   fun["check_partial_equivalence"] = functionality.checkPartialEquivalence;
+  fun["check_approximate_equivalence"] =
+      functionality.checkApproximateEquivalence;
 
   auto& sim = config["simulation"];
   sim["fidelity_threshold"] = simulation.fidelityThreshold;